Add support for input file suffixes
Day 8 provides different examples for the two parts.
This seemed like a useful concept to generalise.
By adding an additional optional parameter to the arguments list,
suffixes can be specified for either example or real inputs.

One limitation with the new implementation is the suffix,
if provided,
is not optional.
One possible future extension would be to make the suffix optional,
in the same way that the exercise label is optional.