Solve day 8 problems
Day 8 part 2 brings with it a quandary.
To answer the problem,
a number of assumptions have to be made about the input data.
(A brute force approach will simply take too long.)
However, it is relatively simple to come up with datasets
that have no valid solution according to the problem statement.