Switch to "text coordinates"
The initial implementation of the day 3 problems relied on mentally
translating x-coordinates to the line in the input,
and the y-coordinate to the character in the line.
The new implementations make that more explicit.

In addition, some of the coordinate methods have been extracted into
library classes to make the more reusable.