Enable log verbosity levels
When adding lots of logging to a particular solution,
not all levels of detail are equally valuable.
Separating out verbosity levels of debug logs
enables discrete levels of logging for additional output,
without necessitating a binary on/off decision for all cases.

So far, only one set of solutions has been refactored
to take advantage of this feature.