Switch to custom loggers
Using the root logger works fine,
especially for such small encapsulated scripts.
But with an increased reliance on complex "visual" output
that we do not even want to compile unless we intend to log it,
it is useful to have loggers to access their current log level.
So this commit switches all scripts to using a logger
instead of the implicit root logger
to conform to that best practice.
The use of the filename in the logger name is a nice addition
in the case that output from multiple scripts is amalgamated.