Support variant scripts
This library change supports auto-searching for relevant files
even when the script is a "variant"
and not one of the primary "solvers".
For the purposes of considering the name of the file,
it makes the assumption that anything after the first hyphen can be ignored.
This means, for example,
that "brute force" variants can be run directly
in the same way as the primary "solvers".